pdf-accounting-importer/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

openai==1.40.6

pymupdf==1.24.9
pdfplumber==0.11.4
pillow==10.4.0
numpy==1.26.4
